requirements.txt
django
google-generativeai